Addition of LF features (#33)
* Added aperiodic and residual lf features

* Version bumps

* Added the specparam to dependecies for additional lf features

* Added utlilty functions to upload and downloads features data from AWS

* Added utility functions for listing latetst labels and downloading features with backward compatilbility.

* Added the Changelog

* Backward compatibility of the outside column

Updated the calculation of the outside column in features dataframe. Now it allows the use of both labels(old version) and channel_labels for the creation of df_features.

* Added list_available_projects function

Created a base class which gets the list of immediate children for a prefix, and then two subclasses which takes it's output to get the latest labels and latest projects.